hft/requirements.txt
# Core data
numpy>=1.26.0
pandas>=2.1.0
pyarrow>=14.0.0
scipy>=1.11.0

# Config and env
python-dotenv>=1.0.0

# HTTP and async (for live execution layer)
aiohttp>=3.9.0
websockets>=12.0
requests>=2.31.0

# AWS (EC2 deployment)
boto3>=1.34.0

# Crypto exchange (Bitso REST API)
# No official SDK — we use requests directly

# Monitoring and alerts
# telegram notifications via requests (no extra dep needed)

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0

# Dev tools
ipykernel>=6.29.0
jupyterlab>=4.0.0
matplotlib>=3.8.0
seaborn>=0.13.0
